move general popup in popups directory
